+++ Lip.Core/GoModuleArchiveFileSource.cs
@@ -11,18 +11,26 @@ public class GoModuleArchiveFileSource(
     SemVersion version) : ArchiveFileSource(fileSystem, archiveFilePath)
 {
     private readonly string _archiveFilePath = archiveFilePath;
+
     private readonly IFileSystem _fileSystem = fileSystem;
-    private readonly string _prefix = $"{goModulePath}@{Module.CanonicalVersion("v" + version.ToString())}/";
+
+    // When major >= 2 and there's no go.mod, GoProxy will add +incompatible in version
+    // Reference: https://stackoverflow.com/questions/57355929/what-does-incompatible-in-go-mod-mean-will-it-cause-harm
+    private readonly string _prefix =
+        $"{goModulePath}@{Module.CanonicalVersion("v" + (version.MetadataIdentifiers.Count == 0 && version.Major >= 2 ? version + "+incompatible" : version.ToString()))}/";
 
     public override async Task<List<IFileSourceEntry>> GetAllEntries()
     {
-        List<IFileSourceEntry> entries = [.. (await base.GetAllEntries())
+        List<IFileSourceEntry> entries =
+        [
+            .. (await base.GetAllEntries())
             .Where(entry => entry.Key.StartsWith(_prefix))
             .Select(entry => new GoModuleArchiveFileSourceEntry(
                 _fileSystem,
                 _archiveFilePath,
                 entry.Key[_prefix.Length..],
-                entry.Key))];
+                entry.Key))
+        ];
 
         return entries;
     }